Remove support for application-supplied config files.
They seemed to be causing more problems than they solved, and none of the
built-in ones provided more than a few lines of logic.

This also removes the default [program:runserver] section, for the same
reasons.  Apart from the auto-reloader, there are now to programs that will
run without you explicitly configuring them to do so.